mattdrayer/api-proforma-to-gradebook: Calculate/store proforma grade in gradebook
