+++ lms/djangoapps/gradebook/receivers.py
@@ -24,10 +24,12 @@ def on_score_changed(sender, **kwargs):
     request.user = user
     grade_data = grades.grade(user, request, course_descriptor)
     grade = grade_data['percent']
+    proforma_grade = grades.calculate_proforma_grade(grade_data, course_descriptor.grading_policy)
     try:
         gradebook_entry = StudentGradebook.objects.get(user=user, course_id=course_key)
         if gradebook_entry.grade != grade:
             gradebook_entry.grade = grade
+            gradebook_entry.proforma_grade = proforma_grade
             gradebook_entry.save()
     except StudentGradebook.DoesNotExist:
-        StudentGradebook.objects.create(user=user, course_id=course_key, grade=grade)
+        StudentGradebook.objects.create(user=user, course_id=course_key, grade=grade, proforma_grade=proforma_grade)
